Renamed ContactNote to AnnonseNote
Renamed ContactNote since the name is no longer sufficient to
describe its function. It is now thought of as a general note, that
can be used and attached in most cases. Modified models to now
link to notes linked to them. Contact as a holder for persons and
notes are now disbanded.